✨ [Refactor] move QuoteProgressContext to a new directory and update import paths
